+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.4.0] - 2026-07-26
+
+### Added
+
+- Added new config options for the search widget:
+    - **Horizontal/vertical padding**
+    - **Vertical offset**
+    - **Align with bottom of panel**
+    - **Text color**
+- Added a config option to change the number of Creative tabs displayed.
+
+### Changed
+
+- The search widget now uses its own unique texture.
+
 ## [3.3.2] - 2026-07-25
 
 ### Fixed